CassandraTimestampUtils.
